chore: Re-enable proxy test
