+++ test-packages/integration-tests/package.json
@@ -21,6 +21,8 @@
   },
   "devDependencies": {
     "@types/jest": "^25.1.4",
+    "@types/http-proxy": "^1.17.4",
+    "http-proxy": "^1.18.0",
     "jest": "^25.1.0",
     "jest-junit": "^10.0.0",
     "mockserver-node": "5.8.1",
