+++ test-packages/integration-tests/test/proxy.spec.ts
@@ -1,206 +1,77 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
+
+import * as http from 'http';
+import * as https from 'https';
+import * as fs from 'fs';
+import * as path from 'path';
+import { once } from 'events';
 import {
-  Destination,
+  Protocol,
   executeHttpRequest,
   HttpMethod,
-  HttpResponse,
-  Protocol
+  Destination
 } from '@sap-cloud-sdk/core';
-import {
-  setTestDestination,
-  unmockTestDestination
-} from '@sap-cloud-sdk/test-util';
-import { MapType } from '@sap-cloud-sdk/util';
-import axios from 'axios';
-import {
-  ODataRequest,
-  ODataRequestConfig
-} from '../../../packages/core/src/request-builder/request';
-import mockserver = require('mockserver-node');
+import httpProxy from 'http-proxy';
 
+const httpServerResponse = 'http';
 describe('proxy', () => {
-  const proxyPort = 1080;
-  const destination: Destination = {
-    name: 'exampleDotCom',
-    url: 'http://example.com',
-    proxyConfiguration: {
-      host: 'localhost',
-      protocol: Protocol.HTTP,
-      port: proxyPort,
-      headers: undefined
-    }
-  };
+  it("should use the http proxy with http-execute for 'http'", async () => {
+    const httpServer = await createHttpServer();
+    const httpServerPort = getServerPort(httpServer);
 
-  beforeAll(async () => {
-    await startProxyServer(proxyPort);
-  }, 10000);
+    const proxySpy = jest.fn();
 
-  afterAll(async () => {
-    await mockserver.stop_mockserver({ serverPort: proxyPort });
-    unmockTestDestination(destination.name as string);
-    delete process.env.VCAP_SERVICES;
-  });
+    const proxyServer = await createProxyServer(proxySpy);
+    const proxyPort = getServerPort(proxyServer);
 
-  beforeEach(async () => {
-    const response = await axios.put(
-      'http://localhost:1080/mockserver/clear?type=log'
-    );
-    expect(response.status).toBe(200);
-    unmockTestDestination(destination.name as string);
-    delete process.env.VCAP_SERVICES;
-  });
-
-  it('should use the http proxy with odata-request', async () => {
-    await checkProxyServerLogsClean();
-    const config = new DummyODataRequestConfig('get', '', 'application/json');
-    const oDataRequest = new ODataRequest(config, destination);
-    const response = await oDataRequest.execute();
-
-    checkResponse(response);
-
-    await checkProxyServerWasUsed();
-  });
-
-  it('should use the http proxy with http-execute.', async () => {
-    await checkProxyServerLogsClean();
-
-    await executeHttpRequest(destination, {
-      method: HttpMethod.GET
-    });
-
-    await checkProxyServerWasUsed();
-  }, 10000);
-
-  it('should use proxy if destination is in env variables.', async () => {
-    setTestDestination(destination);
-
-    await checkProxyServerLogsClean();
-
-    // Do the call using the proxy
-    const response = await executeHttpRequest(
-      { destinationName: destination.name as string },
-      {
-        method: HttpMethod.GET
+    const destination: Destination = {
+      url: `http://localhost:${httpServerPort}`,
+      proxyConfiguration: {
+        host: 'localhost',
+        protocol: Protocol.HTTP,
+        port: proxyPort
       }
-    );
-
-    checkResponse(response);
-
-    await checkProxyServerWasUsed();
-  }, 10000);
-
-  it('should use proxy if destination and proxy is in env variables.', async () => {
-    setTestDestination({
-      name: 'exampleDotCom',
-      url: 'http://example.com'
-    });
-
-    process.env['http_proxy'] = `http://localhost:${proxyPort}`;
-
-    await checkProxyServerLogsClean();
-
-    // Do the call using the proxy
-    const response = await executeHttpRequest(
-      { destinationName: destination.name as string },
-      {
-        method: HttpMethod.GET
-      }
-    );
-
-    checkResponse(response);
-
-    await checkProxyServerWasUsed();
-  }, 10000);
-
-  it('should use proxy if destination is in vcap and proxy is in env variables.', async () => {
-    const serviceBindings = {
-      's4-hana-cloud': [
-        {
-          binding_name: null,
-          credentials: {
-            Authentication: 'BasicAuthentication',
-            Password: 'bar',
-            URL: destination.url,
-            User: 'foo'
-          },
-          instance_name: destination.name,
-          label: 's4-hana-cloud',
-          name: destination.name,
-          plan: 'sap_com_0008',
-          provider: null,
-          syslog_drain_url: null,
-          tags: ['s4-hana-cloud'],
-          volume_mounts: []
-        }
-      ]
     };
 
-    process.env.VCAP_SERVICES = JSON.stringify(serviceBindings);
+    const request = executeHttpRequest(destination, {
+      method: HttpMethod.GET
+    }).then(res => res.data);
 
-    process.env['http_proxy'] = `http://localhost:${proxyPort}`;
+    await expect(request).resolves.toEqual(httpServerResponse);
+    expect(proxySpy).toHaveBeenCalled();
 
-    await checkProxyServerLogsClean();
-
-    // Do the call using the proxy
-    const response = await executeHttpRequest(
-      { destinationName: destination.name as string },
-      {
-        method: HttpMethod.GET
-      }
-    );
-
-    checkResponse(response);
-
-    await checkProxyServerWasUsed();
-  }, 10000);
+    await closeServer(httpServer);
+    await closeServer(proxyServer);
+  });
 });
 
-const expectedStringsAfterRequest = [
-  'example.com',
-  'for forwarded request',
-  'This domain is for use in illustrative examples in documents.'
-];
-
-async function checkProxyServerLogsClean() {
-  const serverLogsBeforCall = await getMockServerLogs();
-  expectedStringsAfterRequest.forEach(expected =>
-    expect(serverLogsBeforCall).not.toContain(expected)
-  );
-}
-
-async function checkResponse(response: HttpResponse) {
-  expect(response.status).toBe(200);
-  expect(response.data).toContain(
-    'This domain is for use in illustrative examples in documents.'
-  );
-}
-
-async function checkProxyServerWasUsed() {
-  const serverLogsAfterCall = await getMockServerLogs();
-  expectedStringsAfterRequest.forEach(expected =>
-    expect(serverLogsAfterCall).toContain(expected)
-  );
-}
-
-async function getMockServerLogs(): Promise<string> {
-  return (
-    await axios.put('http://localhost:1080/mockserver/retrieve?type=logs')
-  ).data;
-}
-
-async function startProxyServer(port: number) {
-  return mockserver.start_mockserver({
-    serverPort: port,
-    verbose: true
+async function createHttpServer() {
+  const server = http.createServer((req, res) => {
+    res.end(httpServerResponse);
   });
+
+  server.listen();
+  await once(server, 'listening');
+  return server;
 }
 
-class DummyODataRequestConfig extends ODataRequestConfig {
-  queryParameters(): MapType<any> {
-    return new Map();
-  }
+async function createProxyServer(proxySpy) {
+  const proxy = httpProxy.createProxyServer({});
+  const server = http.createServer(function (req, res) {
+    proxySpy();
+    proxy.web(req, res, { target: `http://${req.headers.host!}` });
+  });
 
-  resourcePath(): string {
-    return '';
-  }
+  server.listen();
+  await once(server, 'listening');
+  return server;
+}
+
+function getServerPort(server) {
+  return (server._server || server).address().port;
+}
+
+async function closeServer(server) {
+  server.close();
+  return once(server, 'close');
 }
